Front management, gestion des utilisateurs

+++ georchestra-signalement-service/src/main/java/org/georchestra/signalement/service/sm/impl/InitializationServiceImpl.java
@@ -1,12 +1,7 @@
 /**
- * 
+ *
  */
 package org.georchestra.signalement.service.sm.impl;
-
-import java.io.FileInputStream;
-import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.List;
 
 import org.activiti.engine.ProcessEngine;
 import org.activiti.engine.RepositoryService;
@@ -16,7 +11,11 @@ import org.activiti.engine.repository.ProcessDefinitionQuery;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.georchestra.signalement.core.common.DocumentContent;
+import org.georchestra.signalement.core.dto.ContextDescription;
+import org.georchestra.signalement.core.dto.ContextDescriptionSearchCriteria;
+import org.georchestra.signalement.core.dto.SortCriteria;
 import org.georchestra.signalement.service.exception.InitializationException;
+import org.georchestra.signalement.service.exception.InvalidDataException;
 import org.georchestra.signalement.service.mapper.workflow.ProcessDefinitionMapper;
 import org.georchestra.signalement.service.sm.InitializationService;
 import org.slf4j.Logger;
@@ -25,9 +24,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.util.MimeTypeUtils;
 
+import java.io.FileInputStream;
+import java.util.*;
+
 /**
  * @author FNI18300
- *
  */
 @Component
 public class InitializationServiceImpl implements InitializationService {
@@ -36,9 +37,12 @@ public class InitializationServiceImpl implements InitializationService {
 
 	@Autowired
 	private ProcessEngine processEngine;
-	
+
 	@Autowired
 	private ProcessDefinitionMapper processDefinitionMapper;
+
+	@Autowired
+	private ContextDescriptionServiceImpl contextService;
 
 	@Override
 	public void initialize() throws InitializationException {
@@ -97,26 +101,52 @@ public class InitializationServiceImpl implements InitializationService {
 	}
 
 	@Override
-	public boolean deleteProcessDefinition(String processDefinitionName, Integer version) throws InitializationException {
-		boolean result = false;
+	public boolean deleteProcessDefinition(String processDefinitionName, Integer version) throws InvalidDataException {
+		boolean result = true;
 		RepositoryService repositoryService = processEngine.getRepositoryService();
 		ProcessDefinitionQuery query = repositoryService.createProcessDefinitionQuery()
 				.processDefinitionName(processDefinitionName);
-		if(version != null){
-			LOGGER.info("A version has been provided : ", version);
+		if (version != null) {
+			LOGGER.info("A version has been provided : {}", version);
 			query = query.processDefinitionVersion(version);
 		}
 		List<ProcessDefinition> processDefinitions = query.list();
 		if (CollectionUtils.isNotEmpty(processDefinitions)) {
+			Map<String, Set<Integer>> usedWorflows = getUsedWorkflows();
 			for (ProcessDefinition processDefinition : processDefinitions) {
-				LOGGER.info("Start delete deployment {}.", processDefinition.getDeploymentId());
-				repositoryService.deleteDeployment(processDefinition.getDeploymentId(), true);
-				LOGGER.info("Delate deployment {} done.", processDefinition.getDeploymentId());
+				if (!processDefinitionIsUsed(processDefinition, usedWorflows)) {
+					LOGGER.info("Start delete deployment {}.", processDefinition.getDeploymentId());
+					repositoryService.deleteDeployment(processDefinition.getDeploymentId(), true);
+					LOGGER.info("Delate deployment {} done.", processDefinition.getDeploymentId());
+				} else {
+					LOGGER.info("Can't delete used workflow : {}.", processDefinition.getDeploymentId());
+					result = false;
+				}
 			}
-			result = true;
+		} else {
+			String msg = processDefinitionName + " not found";
+			throw new InvalidDataException(msg);
 		}
 		return result;
 	}
 
+	private Map<String, Set<Integer>> getUsedWorkflows() {
+		ContextDescriptionSearchCriteria searchCriteria = new ContextDescriptionSearchCriteria();
+		SortCriteria sortCriteria = new SortCriteria();
+		List<ContextDescription> contexts = contextService.searchContextDescriptions(searchCriteria, sortCriteria);
+		Map<String, Set<Integer>> usedWorkflows = new HashMap<>();
+		for (ContextDescription context : contexts) {
+			if (!usedWorkflows.containsKey(context.getProcessDefinitionKey())) {
+				usedWorkflows.put(context.getProcessDefinitionKey(), new HashSet<>());
+			}
+			usedWorkflows.get(context.getProcessDefinitionKey()).add(context.getRevision());
+		}
+		return usedWorkflows;
+	}
 
+	private boolean processDefinitionIsUsed(ProcessDefinition processDefinition, Map<String, Set<Integer>> usedWorflows) {
+		String key = processDefinition.getKey();
+		Integer revision = processDefinition.getVersion();
+		return usedWorflows.containsKey(key) && usedWorflows.get(key).contains(revision);
+	}
 }
